django/requirements.txt
Django>=4.2.4
ipython>=8.14.0
uvicorn>=0.23.2
whitenoise>=6.5.0
django-ninja>=0.22.2
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
ipython>=8.14.0
orjson>=3.9.4
whitenoise>=6.5.0